feat: Add name input box to registration form
This commit adds a new input box for the user's name to the registration form in the `register1.blade.php` file. The input box is labeled "Name" and is of type "string". This change allows users to enter their name during the registration process.

Note: This commit message follows the established convention of starting with a type (feat for feature) and providing a concise and descriptive summary of the changes made.