feat: Add user dropdown to course registration form
The code changes in the `CourseController.php` file add functionality to retrieve and display a list of users in a dropdown menu on the course registration form. This is achieved by importing the `users` model and fetching all users from the database. The retrieved users are then passed to the `courseadd.blade.php` view, where they are used to populate the dropdown menu for selecting a user.

Note: This commit message follows the established convention of starting with a type (feat for feature) and providing a concise and descriptive summary of the changes made.