chore: Update link for viewing institutes in index.blade.php
This commit updates the link for viewing institutes in the index.blade.php file. The previous link was pointing to "/institute" but it has been changed to "/institute/instituteshow". This change ensures that the correct route is used for displaying the list of institutes.

Note: This commit message follows the established convention of starting with a type (chore for miscellaneous tasks) and providing a concise and descriptive summary of the changes made.